Move some data structures within `sc-consensus-subspace` into their corresponding modules
