+++ crates/subspace-node/src/domain/domain_instance_starter.rs
@@ -10,8 +10,9 @@ use domain_service::{FullBackend, FullClient};
 use futures::StreamExt;
 use sc_chain_spec::ChainSpec;
 use sc_cli::{CliConfiguration, Database, DefaultConfigurationValues, SubstrateCli};
+use sc_consensus_subspace::block_import::BlockImportingNotification;
 use sc_consensus_subspace::notification::SubspaceNotificationStream;
-use sc_consensus_subspace::{BlockImportingNotification, NewSlotNotification};
+use sc_consensus_subspace::slot_worker::NewSlotNotification;
 use sc_service::{BasePath, Configuration};
 use sc_transaction_pool_api::OffchainTransactionPoolFactory;
 use sc_utils::mpsc::{TracingUnboundedReceiver, TracingUnboundedSender};
