+++ crates/subspace-service/src/rpc.rs
@@ -24,11 +24,10 @@
 use jsonrpsee::RpcModule;
 use pallet_transaction_payment_rpc::{TransactionPayment, TransactionPaymentApiServer};
 use sc_client_api::{AuxStore, BlockBackend};
-use sc_consensus_subspace::archiver::SegmentHeadersStore;
+use sc_consensus_subspace::archiver::{ArchivedSegmentNotification, SegmentHeadersStore};
 use sc_consensus_subspace::notification::SubspaceNotificationStream;
-use sc_consensus_subspace::slot_worker::SubspaceSyncOracle;
-use sc_consensus_subspace::{
-    ArchivedSegmentNotification, NewSlotNotification, RewardSigningNotification,
+use sc_consensus_subspace::slot_worker::{
+    NewSlotNotification, RewardSigningNotification, SubspaceSyncOracle,
 };
 use sc_consensus_subspace_rpc::{SubspaceRpc, SubspaceRpcApiServer, SubspaceRpcConfig};
 use sc_rpc::SubscriptionTaskExecutor;
